refactor: moved yap service helpers to its own file and clean up yap handler
